requirements.txt
numpy==1.21.1
opencv-python==4.5.3.56
Pillow==8.3.1
torch==1.9.0+cu102
torchaudio==0.9.0
torchvision==0.10.0+cu102
typing-extensions==3.10.0.0
